Use a non-mounted directory for nodejs data
Set the FLOW_NODEJS_DIR variable in the docker image so that it doesn't
use a directory under /workspace. Since /workspace is mounted,
reading/writing files there is very slow on mac, where they need to be
synced between the OSX host and the linux VM. This made nodejs builds
extremely slow, and moving it seems to cut roughly a minute off of full
build times.